refactor: restructure Python project into src/index_ripper package
- Move all source modules into src/index_ripper/ with proper package structure
- Split ui_ctk.py (1500 LOC) into app.py, ui/filetree.py, ui/filters.py
- Move tests into tests/ directory, consolidate test_app_utils + test_security_utils
- Remove legacy code: index_ripper.py (old UI), ui_ttkb.py, pyside_poc.py, POC_MATRIX.md
- Remove ttkbootstrap dependency
- Add hatchling build-system for src layout
- Update CI, PyInstaller, and README for new structure
- Fix: duplicate macOS keyboard bindings, expand folders by default
- All 42 tests pass, smoke test and self-test verified
