--- index_ripper.py
@@ -1,1434 +1,0 @@
-"""Index Ripper UI using customtkinter with search, logs, and per-file downloads."""
-
-# pylint: disable=too-many-lines
-import os
-import posixpath
-import sys
-import threading
-from concurrent.futures import ThreadPoolExecutor
-from queue import Empty, Queue
-from threading import Event, Thread
-from urllib.parse import unquote, urljoin, urlparse
-
-if "--smoke" in sys.argv:
-    print("SMOKE_OK")
-    raise SystemExit(0)
-
-if "--self-test" in sys.argv:
-    from self_test import run_self_test
-
-    result = run_self_test()
-    print(
-        f"SELF_TEST_OK total={result.total} files={result.files} dirs={result.directories}"
-    )
-    raise SystemExit(0)
-
-
-from app_utils import configure_tk_libraries
-configure_tk_libraries()
-
-import tkinter as tk
-from tkinter import filedialog, messagebox, ttk
-
-import customtkinter as ctk  # type: ignore
-import requests  # type: ignore
-from requests.adapters import HTTPAdapter  # type: ignore
-from urllib3.util.retry import Retry  # type: ignore
-
-from backend import Backend
-from ui_downloads import DownloadsPanel
-from app_utils import (
-    build_download_path,
-    default_download_folder,
-    normalize_extension,
-    rebuild_executor,
-    safe_join,
-    sanitize_filename,
-    sanitize_path_segment,
-    shorten_path,
-)
-from settings_store import load_settings, save_settings
-from ui_theme import (
-    action_button_style_name,
-    apply_app_theme,
-    configure_action_button_styles,
-    configure_treeview_style,
-    treeview_tag_colors,
-    ui_tokens,
-)
-
-
-class WebsiteCopier:
-    """A GUI application for downloading files from a website directory listing."""
-
-    USER_AGENT = (
-        "Mozilla/5.0 (Windows NT 10.0; Win64; x64) "
-        "AppleWebKit/537.36 (KHTML, like Gecko) "
-        "Chrome/91.0.4472.124 Safari/537.36"
-    )
-
-    def __init__(self):
-        """Initializes the main window and UI components."""
-        apply_app_theme(ctk)
-
-        self.window = ctk.CTk()
-        self.window.title("Index Ripper")
-        self.window.geometry("1200x900")
-        self.window.minsize(900, 650)
-        configure_action_button_styles(self.window, ctk, ttk)
-        self.ui_tokens = ui_tokens()
-
-        self.backend = Backend(self)
-
-        # Download control
-        self.pause_event = Event()
-        self.current_downloads = []
-
-        # Add scan control
-        self.scan_pause_event = Event()
-        self.scan_pause_event.set()  # Initially set to not paused
-        self.is_scanning = False
-
-        # 添加线程锁，用于保护共享数据
-        self.files_dict_lock = threading.Lock()
-        self.folders_dict_lock = threading.Lock()
-
-        # 初始化队列和处理状态
-        self.dir_queue = Queue()
-        self.file_queue = Queue()
-        self.scan_item_buffer = Queue()
-        self.is_processing_dirs = False
-        self.is_processing_files = False
-        self.scan_flush_interval_ms = 16
-        self.scan_flush_batch_size = 200
-        self.scan_flush_job = None
-
-        # --- Main Layout ---
-        self.window.grid_columnconfigure(0, weight=1)
-        self.window.grid_rowconfigure(1, weight=1)
-
-        # --- URL and filter area ---
-        self.top_frame = ctk.CTkFrame(self.window, corner_radius=14, border_width=1)
-        self.top_frame.grid(row=0, column=0, padx=10, pady=(10, 5), sticky="ew")
-        self.top_frame.grid_columnconfigure(0, weight=1)
-        ctk.CTkLabel(
-            self.top_frame,
-            text="Index Ripper Studio",
-            font=ctk.CTkFont(size=20, weight="bold"),
-        ).grid(row=0, column=0, padx=12, pady=(10, 2), sticky="w")
-
-        # URL input
-        url_input_frame = ctk.CTkFrame(self.top_frame, fg_color="transparent")
-        url_input_frame.grid(row=1, column=0, padx=10, pady=5, sticky="ew")
-        url_input_frame.grid_columnconfigure(1, weight=1)
-        ctk.CTkLabel(
-            url_input_frame, text="URL:", font=ctk.CTkFont(weight="bold")
-        ).grid(row=0, column=0, padx=(0, 5))
-        self.url_entry = ctk.CTkEntry(
-            url_input_frame, placeholder_text="https://example.com", height=36
-        )
-        self.url_entry.grid(row=0, column=1, sticky="ew")
-
-        # Scan buttons
-        scan_buttons_frame = ctk.CTkFrame(self.top_frame, fg_color="transparent")
-        scan_buttons_frame.grid(row=1, column=1, padx=10, pady=5)
-
-        self.status_chip = ctk.CTkLabel(
-            self.top_frame,
-            text="Ready",
-            fg_color=("#DCFCE7", "#14532D"),
-            text_color=("#14532D", "#DCFCE7"),
-            corner_radius=12,
-            padx=10,
-            pady=4,
-        )
-        self.status_chip.grid(row=0, column=2, padx=(0, 10), pady=8, sticky="e")
-
-        self.scan_btn = ttk.Button(
-            scan_buttons_frame,
-            text="Scan",
-            command=self.start_scan,
-            style=action_button_style_name("primary"),
-        )
-        self.scan_btn.pack(side="left", padx=5)
-
-        self.scan_pause_btn = ttk.Button(
-            scan_buttons_frame,
-            text="Pause Scan",
-            command=self.toggle_scan_pause,
-            state="disabled",
-            style=action_button_style_name("secondary"),
-        )
-        self.scan_pause_btn.pack(side="left", padx=5)
-
-        self.clear_scan_btn = ttk.Button(
-            scan_buttons_frame,
-            text="Clear",
-            command=self.clear_scan_results,
-            style=action_button_style_name("danger"),
-        )
-        self.clear_scan_btn.pack(side="left", padx=5)
-
-        # --- File type filter area ---
-        filter_frame = ctk.CTkFrame(self.top_frame, corner_radius=10)
-        filter_frame.grid(row=2, column=0, columnspan=2, padx=10, pady=5, sticky="ew")
-        filter_frame.grid_columnconfigure(0, weight=1)
-
-        ctk.CTkLabel(
-            filter_frame, text="File Type Selection", font=ctk.CTkFont(weight="bold")
-        ).grid(row=0, column=0, padx=10, pady=(5, 0), sticky="w")
-
-        # File type checkbox container
-        self.filter_checkboxes_frame = ctk.CTkScrollableFrame(
-            filter_frame, label_text="", fg_color="transparent"
-        )
-        self.filter_checkboxes_frame.grid(row=1, column=0, padx=5, pady=5, sticky="ew")
-
-        # Variable to store file types
-        self.file_types = {}  # {'.pdf': BooleanVar(), '.jpg': BooleanVar(), ...}
-        self.file_type_counts = {}  # {'.pdf': 0, '.jpg': 0, ...}
-
-        # Select/deselect all buttons
-        select_buttons_frame = ctk.CTkFrame(filter_frame, fg_color="transparent")
-        select_buttons_frame.grid(row=2, column=0, padx=10, pady=5, sticky="w")
-        ctk.CTkButton(
-            select_buttons_frame, text="Select All", command=self.select_all_types
-        ).pack(side="left", padx=2)
-        ctk.CTkButton(
-            select_buttons_frame, text="Deselect All", command=self.deselect_all_types
-        ).pack(side="left", padx=2)
-
-        # --- File list area ---
-        self.tree_frame = ctk.CTkFrame(self.window, corner_radius=14, border_width=1)
-        self.tree_frame.grid(row=1, column=0, padx=10, pady=5, sticky="nsew")
-        self.tree_frame.grid_rowconfigure(1, weight=1)
-        self.tree_frame.grid_columnconfigure(0, weight=1)
-
-        # Search bar
-        search_frame = ctk.CTkFrame(self.tree_frame)
-        search_frame.grid(row=0, column=0, columnspan=2, padx=5, pady=5, sticky="ew")
-        search_frame.grid_columnconfigure(1, weight=1)
-
-        ctk.CTkLabel(search_frame, text="Search:").grid(row=0, column=0, padx=(5, 2))
-        self.search_var = ctk.StringVar()
-        self.search_var.trace_add("write", self.on_search_filter_changed)
-        self.search_entry = ctk.CTkEntry(
-            search_frame,
-            textvariable=self.search_var,
-            state="disabled",
-            placeholder_text="Filter by file or folder name...",
-            height=32,
-        )
-        self.search_entry.grid(row=0, column=1, sticky="ew")
-        self.full_tree_backup = {}
-
-        configure_treeview_style(self.window, ctk, ttk)
-
-        self.tree = ttk.Treeview(
-            self.tree_frame,
-            selectmode="browse",
-            show=("tree", "headings"),
-            style="Treeview",
-        )
-
-        colors = treeview_tag_colors(self.window)
-        self.tree.tag_configure("oddrow", background=colors["oddrow"])
-        self.tree.tag_configure("evenrow", background=colors["evenrow"])
-        self.tree.tag_configure("checked", foreground=colors["checked"])
-
-        # Set column configuration
-        self.tree["columns"] = ("size", "type")
-        self.tree.heading("#0", text="Name", command=lambda: self.sort_tree("name"))
-        self.tree.heading("size", text="Size", command=lambda: self.sort_tree("size"))
-        self.tree.heading("type", text="Type", command=lambda: self.sort_tree("type"))
-
-        self.tree.column("#0", minwidth=300, width=600, stretch=True)
-        self.tree.column("size", width=120, anchor="center")
-        self.tree.column("type", width=200, anchor="center")
-
-        # Bind right-click menu
-        self.tree.bind("<Button-3>", self.show_context_menu)
-
-        # Create right-click menu (using tk.Menu as customtkinter doesn't have one)
-        self.context_menu = tk.Menu(self.window, tearoff=0)
-        self.context_menu.add_command(label="Select All", command=self.select_all)
-        self.context_menu.add_command(label="Deselect All", command=self.deselect_all)
-        self.context_menu.add_separator()
-        self.context_menu.add_command(label="Expand All", command=self.expand_all)
-        self.context_menu.add_command(label="Collapse All", command=self.collapse_all)
-
-        # Sort status
-        self.sort_column = None
-        self.sort_reverse = False
-
-        self.checked_items = set()
-        self.checkbox_checked = "✔"
-        self.folder_icon = "📁"
-        self.file_icon = "📄"
-        self.folders = {}
-
-        self.scrollbar = ctk.CTkScrollbar(self.tree_frame, command=self.tree.yview)
-        self.h_scrollbar = ctk.CTkScrollbar(
-            self.tree_frame, orientation="horizontal", command=self.tree.xview
-        )
-        self.tree.configure(
-            yscrollcommand=self.scrollbar.set, xscrollcommand=self.h_scrollbar.set
-        )
-
-        self.tree.grid(row=1, column=0, sticky="nsew")
-        self.scrollbar.grid(row=1, column=1, sticky="ns")
-        self.h_scrollbar.grid(row=2, column=0, sticky="ew")
-
-        self.tree.bind("<Button-1>", self.on_tree_click)
-        self.tree.bind("<space>", self.on_tree_space)
-        self.tree.bind("<Return>", self.on_tree_enter)
-        self.tree.bind("<Control-a>", self.on_tree_select_all)
-        self.tree.bind("<Command-a>", self.on_tree_select_all)
-
-        # --- Download control area ---
-        control_frame = ctk.CTkFrame(self.window, corner_radius=14, border_width=1)
-        control_frame.grid(row=2, column=0, padx=10, pady=5, sticky="ew")
-
-        self.download_btn = ttk.Button(
-            control_frame,
-            text="Download Selected Files",
-            command=self.download_selected,
-            style=action_button_style_name("success"),
-        )
-        self.download_btn.pack(side="left", padx=5, pady=5)
-
-        self.pause_btn = ttk.Button(
-            control_frame,
-            text="Pause Download",
-            command=self.toggle_pause,
-            state="disabled",
-            style=action_button_style_name("secondary"),
-        )
-        self.pause_btn.pack(side="left", padx=5, pady=5)
-
-        self.path_btn = ttk.Button(
-            control_frame,
-            text="Choose Download Location",
-            command=self.choose_download_path,
-            style=action_button_style_name("secondary"),
-        )
-        self.path_btn.pack(side="left", padx=5, pady=5)
-
-        # Thread download count control
-        threads_frame = ctk.CTkFrame(control_frame, fg_color="transparent")
-        threads_frame.pack(side="right", padx=5, pady=5)
-        ctk.CTkLabel(threads_frame, text="Concurrent Downloads:").pack(side="left")
-        self.threads_var = ctk.StringVar(value="5")
-        threads_option_menu = ctk.CTkOptionMenu(
-            threads_frame,
-            values=[str(i) for i in range(1, 11)],
-            variable=self.threads_var,
-            command=self.update_thread_count,
-            **self.ui_tokens["option_menu"],
-        )
-        threads_option_menu.pack(side="left", padx=5)
-
-        # Toggle button to show/hide bottom panels (Downloads/Logs)
-        self.panels_visible = True
-        self.toggle_panels_btn = ttk.Button(
-            control_frame,
-            text="Hide Panels",
-            command=self.toggle_panels,
-            style=action_button_style_name("secondary"),
-        )
-        self.toggle_panels_btn.pack(side="right", padx=5)
-
-        # --- Progress bar ---
-        self.progress_frame = ctk.CTkFrame(self.window, corner_radius=14, border_width=1)
-        self.progress_frame.grid(row=3, column=0, padx=10, pady=(5, 10), sticky="ew")
-        self.progress_frame.grid_columnconfigure(0, weight=1)
-
-        self.progress_var = ctk.DoubleVar()
-        self.progress_bar = ctk.CTkProgressBar(
-            self.progress_frame, variable=self.progress_var
-        )
-        self.progress_bar.set(0)
-        self.progress_bar.grid(row=0, column=0, padx=5, pady=2, sticky="ew")
-
-        self.progress_label = ctk.CTkLabel(self.progress_frame, text="")
-        self.progress_label.grid(row=1, column=0, padx=5, pady=2, sticky="w")
-
-        # --- Panels (Downloads + Logs) inside a Tabview ---
-        self.panels_tabview = ctk.CTkTabview(
-            self.window, corner_radius=14, **self.ui_tokens["tabview"]
-        )
-        self.panels_tabview.grid(row=4, column=0, padx=10, pady=(0, 10), sticky="ew")
-
-        self.panels_tabview.add("Downloads")
-        self.panels_tabview.add("Logs")
-
-        downloads_tab = self.panels_tabview.tab("Downloads")
-        self.downloads_frame = ctk.CTkScrollableFrame(downloads_tab, label_text="")
-        self.downloads_frame.pack(fill="x", expand=False)
-
-        logs_tab = self.panels_tabview.tab("Logs")
-        self.log_text = ctk.CTkTextbox(
-            logs_tab,
-            height=self.ui_tokens["logs"]["height"],
-            font=self.ui_tokens["logs"]["font"],
-            text_color=self.ui_tokens["logs"]["text_color"],
-            fg_color=self.ui_tokens["logs"]["fg_color"],
-        )
-        self.log_text.pack(fill="x", expand=False)
-        self.log_text.configure(state="disabled")
-
-        try:
-            self.panels_tabview.set("Logs")
-        except tk.TclError:
-            pass
-        
-        # Track per-file download UI elements and cancel events
-        self.downloads_panel = DownloadsPanel(
-            self.downloads_frame, ctk, tk, threading, tokens=self.ui_tokens
-        )
-
-        # Load and apply saved settings (panels visibility and active tab)
-        self._settings_path = os.path.join(
-            os.path.expanduser("~"),
-            ".index_ripper_settings.json",
-        )
-        saved = load_settings(self._settings_path) if os.path.exists(self._settings_path) else {}
-        # Active tab
-        active = saved.get("active_tab")
-        if active in ("Downloads", "Logs"):
-            try:
-                self.panels_tabview.set(active)
-            except tk.TclError:
-                pass
-        visible = bool(saved.get("panels_visible", True))
-        if visible:
-            self.panels_tabview.grid()
-            self.panels_visible = True
-            self.toggle_panels_btn.configure(text="Hide Panels")
-        else:
-            self.panels_tabview.grid_remove()
-            self.panels_visible = False
-            self.toggle_panels_btn.configure(text="Show Panels")
-
-        # Initialize session with retry
-        retry_strategy = Retry(
-            total=3,
-            backoff_factor=0.1,
-            status_forcelist=[500, 502, 503, 504],
-        )
-
-        self.files_dict = {}
-        self.download_path = ""
-        self.is_paused = False
-        self.download_queue = Queue()
-        self.max_workers = 5
-        self.executor = ThreadPoolExecutor(max_workers=self.max_workers)
-        self.active_downloads = []
-
-        self.session = requests.Session()
-        adapter = HTTPAdapter(
-            max_retries=retry_strategy,
-            pool_connections=10,
-            pool_maxsize=10,
-            pool_block=False,
-        )
-        self.session.mount("http://", adapter)
-        self.session.mount("https://", adapter)
-        self.timeout = (10, 30)
-
-        self.total_urls = 0
-        self.scanned_urls = 0
-        self.window.protocol("WM_DELETE_WINDOW", self.on_closing)
-        self.should_stop = False
-        self.window.bind("<Control-f>", self.focus_search)
-        self.window.bind("<Control-l>", self.focus_logs)
-        self.window.bind("<Escape>", self.clear_search)
-
-    def notify_info(self, title: str, message: str) -> None:
-        try:
-            messagebox.showinfo(title, message)
-        except Exception:
-            self.log_message(f"[INFO] {title}: {message}")
-
-    def notify_warning(self, title: str, message: str) -> None:
-        try:
-            messagebox.showwarning(title, message)
-        except Exception:
-            self.log_message(f"[WARN] {title}: {message}")
-
-    def notify_error(self, title: str, message: str) -> None:
-        try:
-            messagebox.showerror(title, message)
-        except Exception:
-            self.log_message(f"[ERROR] {title}: {message}")
-
-    def _set_status(self, text: str, state: str = "ready") -> None:
-        """Set status chip text and style based on scan state."""
-        status_styles = {
-            "ready": {
-                "fg_color": ("#DCFCE7", "#14532D"),
-                "text_color": ("#14532D", "#DCFCE7"),
-            },
-            "scanning": {
-                "fg_color": ("#DBEAFE", "#1E3A8A"),
-                "text_color": ("#1E3A8A", "#DBEAFE"),
-            },
-            "paused": {
-                "fg_color": ("#FEF3C7", "#78350F"),
-                "text_color": ("#78350F", "#FEF3C7"),
-            },
-            "stopping": {
-                "fg_color": ("#FEE2E2", "#7F1D1D"),
-                "text_color": ("#7F1D1D", "#FEE2E2"),
-            },
-        }
-
-        style = status_styles.get(state, status_styles["ready"])
-        self.status_chip.configure(
-            text=text,
-            fg_color=style["fg_color"],
-            text_color=style["text_color"],
-        )
-
-    def toggle_panels(self):
-        """Show or hide the bottom panels (Downloads/Logs)."""
-        if self.panels_visible:
-            try:
-                self.panels_tabview.grid_remove()
-            except tk.TclError:
-                pass
-            self.panels_visible = False
-            self.toggle_panels_btn.configure(text="Show Panels")
-        else:
-            try:
-                self.panels_tabview.grid()
-            except tk.TclError:
-                pass
-            self.panels_visible = True
-            self.toggle_panels_btn.configure(text="Hide Panels")
-        # Save setting
-        # Persist setting (best-effort)
-        data = {
-            "panels_visible": bool(self.panels_visible),
-            "active_tab": str(self.panels_tabview.get()),
-        }
-        save_settings(self._settings_path, data)
-
-    def choose_download_path(self):
-        """Opens a dialog to choose a download directory."""
-        path = filedialog.askdirectory(title="Choose Download Location")
-        if path:
-            self.download_path = path
-            self.path_btn.configure(text=f"Location: {shorten_path(path)}")
-
-    def toggle_pause(self):
-        """Pauses or resumes the download process."""
-        if self.is_paused:
-            self.pause_event.set()
-            self.pause_btn.configure(text="Pause Download")
-            self.is_paused = False
-        else:
-            self.pause_event.clear()
-            self.pause_btn.configure(text="Resume Download")
-            self.is_paused = True
-
-    def toggle_scan_pause(self):
-        """Toggle scan pause state"""
-        if self.scan_pause_event.is_set():
-            self.scan_pause_event.clear()
-            self.scan_pause_btn.configure(text="Resume Scan")
-            self.progress_label.configure(text="Scan paused")
-            self._set_status("Scan Paused", "paused")
-        else:
-            self.scan_pause_event.set()
-            self.scan_pause_btn.configure(text="Pause Scan")
-            self._set_status("Scanning", "scanning")
-
-    def filter_file_type(self, file_name):
-        """Check if file should be displayed"""
-        ext = normalize_extension(file_name)
-
-        existing_ext = next(
-            (e for e in self.file_types if e.lower() == ext.lower()), None
-        )
-
-        return existing_ext is not None and self.file_types[existing_ext].get()
-
-    def create_folder_structure(self, path, url):
-        """Create folder structure and return ID of last folder"""
-        path = unquote(path.replace("\\", "/"))
-        path = posixpath.normpath(path).strip("/")
-
-        if not path:
-            return ""
-
-        parts = path.split("/")
-        current_path = ""
-        parent = ""
-
-        for part in parts:
-            if not part:
-                continue
-
-            current_path = posixpath.join(current_path, part) if current_path else part
-
-            with self.folders_dict_lock:
-                if current_path in self.folders:
-                    parent = self.folders[current_path]["id"]
-                    continue
-
-                # Check for existing folder by name under the same parent
-                found = False
-                for child in self.tree.get_children(parent):
-                    child_text = self.tree.item(child)["text"]
-                    child_tags = self.tree.item(child)["tags"]
-                    if "folder" in child_tags:
-                        # Extract folder name (remove icon)
-                        folder_name = child_text.replace(f"{self.folder_icon} ", "")
-                        if folder_name == part:
-                            parent = child
-                            self.folders[current_path] = {
-                                "id": child,
-                                "url": urljoin(url, current_path),
-                                "full_path": current_path,
-                            }
-                            found = True
-                            break
-                if found:
-                    continue
-
-                folder_id = self.tree.insert(
-                    parent,
-                    "end",
-                    text=f"{self.folder_icon} {part}",
-                    values=("", "Directory"),
-                    tags=("folder",),
-                )
-                self.folders[current_path] = {
-                    "id": folder_id,
-                    "url": urljoin(url, current_path),
-                    "full_path": current_path,
-                }
-                parent = folder_id
-        return parent
-
-    def _get_ancestors(self, item):
-        """Get all ancestor nodes of an item"""
-        ancestors = []
-        parent = self.tree.parent(item)
-        while parent:
-            ancestors.append(parent)
-            parent = self.tree.parent(parent)
-        return ancestors
-
-    def on_tree_click(self, event):
-        """Handle tree item click event."""
-        item = self.tree.identify("item", event.x, event.y)
-        if not item:
-            return
-
-        self.tree.focus(item)
-        self.tree.selection_set(item)
-
-        # Identify the specific element clicked
-        element = self.tree.identify_element(event.x, event.y)
-
-        # If the click is on the expand/collapse indicator, let the default
-        # binding handle it and do nothing more.
-        if "indicator" in element:
-            return
-
-        # If the click is on the cell (text/icon), toggle the check state.
-        column = self.tree.identify_column(event.x)
-        if column == "#0":  # Only toggle check if name column is clicked
-            self.toggle_check(item)
-
-    def _focused_tree_item(self):
-        item = self.tree.focus()
-        if item and self.tree.exists(item):
-            return item
-        selected = self.tree.selection()
-        if selected:
-            candidate = selected[0]
-            if self.tree.exists(candidate):
-                return candidate
-        return ""
-
-    def on_tree_space(self, _event=None):
-        """Toggle checked state for the focused tree item."""
-        item = self._focused_tree_item()
-        if item:
-            self.toggle_check(item)
-            return "break"
-        return None
-
-    def on_tree_enter(self, _event=None):
-        """Expand/collapse folders or toggle files for the focused tree item."""
-        item = self._focused_tree_item()
-        if not item:
-            return None
-
-        tags = self.tree.item(item, "tags")
-        if "folder" in tags:
-            self.tree.item(item, open=not bool(self.tree.item(item, "open")))
-        else:
-            self.toggle_check(item)
-        return "break"
-
-    def on_tree_select_all(self, _event=None):
-        """Select all files in tree when tree widget has focus."""
-        self.select_all()
-        return "break"
-
-    def toggle_check(self, item, force_check=None):
-        """Toggle item checked status"""
-        if not self.tree.exists(item):
-            return
-
-        tags = list(self.tree.item(item, "tags"))
-        is_checked = "checked" in tags
-
-        if force_check is not None:
-            new_check_state = force_check
-        else:
-            new_check_state = not is_checked
-
-        if new_check_state:
-            if "checked" not in tags:
-                tags.append("checked")
-            self.checked_items.add(item)
-        else:
-            if "checked" in tags:
-                tags.remove("checked")
-            self.checked_items.discard(item)
-
-        self.tree.item(item, tags=tuple(tags))
-
-        if "folder" in self.tree.item(item, "tags"):
-            for child in self.tree.get_children(item):
-                self.toggle_check(child, force_check=new_check_state)
-
-    def update_scan_progress(self):
-        """Update scan progress"""
-        if self.total_urls > 0:
-            progress = self.scanned_urls / self.total_urls
-            self.progress_bar.set(progress)
-            self.progress_label.configure(
-                text=f"Scanning... ({self.scanned_urls}/{self.total_urls}) {progress:.1%}"
-            )
-            self.window.update_idletasks()
-
-    def _run_on_ui_thread(self, callback, *args, wait=False):
-        """Run a callback on Tk main thread."""
-        if threading.current_thread() is threading.main_thread():
-            callback(*args)
-            return
-
-        done = threading.Event()
-        error_box = {}
-
-        def wrapped():
-            try:
-                callback(*args)
-            except Exception as ex:
-                error_box["error"] = ex
-            finally:
-                done.set()
-
-        self.window.after(0, wrapped)
-        if wait:
-            done.wait()
-            if "error" in error_box:
-                raise error_box["error"]
-
-    def on_scan_started(self, url):
-        """Backend hook: initialize UI state before scan starts."""
-        self._run_on_ui_thread(self._on_scan_started_ui, url, wait=True)
-
-    def _on_scan_started_ui(self, _url):
-        self.is_scanning = True
-        self.scan_pause_btn.configure(state="normal", text="Pause Scan")
-        self.progress_label.configure(text="Scanning website...")
-        self.scan_btn.configure(state="normal", text="Stop Scan")
-        self._set_status("Scanning", "scanning")
-
-        self._cancel_scan_item_flush()
-        while not self.scan_item_buffer.empty():
-            self.scan_item_buffer.get()
-
-        while not self.dir_queue.empty():
-            self.dir_queue.get()
-        while not self.file_queue.empty():
-            self.file_queue.get()
-        self.is_processing_dirs = False
-        self.is_processing_files = False
-
-        with self.files_dict_lock:
-            self.files_dict.clear()
-        with self.folders_dict_lock:
-            self.folders.clear()
-
-        self.tree.delete(*self.tree.get_children())
-        self.checked_items.clear()
-        self.file_types.clear()
-        self.file_type_counts.clear()
-        for widget in self.filter_checkboxes_frame.winfo_children():
-            widget.destroy()
-
-        self.total_urls = 0
-        self.scanned_urls = 0
-        self.progress_bar.set(0)
-        self.search_entry.configure(state="disabled")
-        self.search_var.set("")
-
-    def on_scan_progress(self, scanned_urls, total_urls):
-        """Backend hook: update scan counters/progress."""
-        self._run_on_ui_thread(self._on_scan_progress_ui, scanned_urls, total_urls)
-
-    def _on_scan_progress_ui(self, scanned_urls, total_urls):
-        self.scanned_urls = scanned_urls
-        self.total_urls = total_urls
-        self.update_scan_progress()
-
-    def on_scan_item(
-        self,
-        is_directory,
-        path,
-        url,
-        file_name=None,
-        size=None,
-        file_type=None,
-        full_path=None,
-    ):
-        """Backend hook: enqueue a discovered directory/file for UI processing."""
-        self._run_on_ui_thread(
-            self._on_scan_item_ui,
-            is_directory,
-            path,
-            url,
-            file_name,
-            size,
-            file_type,
-            full_path,
-        )
-
-    def _on_scan_item_ui(
-        self,
-        is_directory,
-        path,
-        url,
-        file_name,
-        size,
-        file_type,
-        full_path,
-    ):
-        if not self.is_scanning:
-            return
-        self.scan_item_buffer.put(
-            (is_directory, path, url, file_name, size, file_type, full_path)
-        )
-        self._schedule_scan_item_flush()
-
-    def _schedule_scan_item_flush(self):
-        if self.scan_flush_job is not None:
-            return
-        if not self.is_scanning:
-            return
-        self.scan_flush_job = self.window.after(
-            self.scan_flush_interval_ms, self._flush_scan_item_buffer
-        )
-
-    def _cancel_scan_item_flush(self):
-        if self.scan_flush_job is None:
-            return
-        try:
-            self.window.after_cancel(self.scan_flush_job)
-        except tk.TclError:
-            pass
-        finally:
-            self.scan_flush_job = None
-
-    def _flush_scan_item_buffer(self, max_items=None, reschedule=True):
-        self.scan_flush_job = None
-        if max_items is None:
-            max_items = self.scan_flush_batch_size
-
-        processed = 0
-        added_dir = False
-        added_file = False
-
-        while processed < max_items:
-            try:
-                (
-                    is_directory,
-                    path,
-                    url,
-                    file_name,
-                    size,
-                    file_type,
-                    full_path,
-                ) = self.scan_item_buffer.get_nowait()
-            except Empty:
-                break
-
-            if is_directory:
-                with self.folders_dict_lock:
-                    self.dir_queue.put((path, url))
-                added_dir = True
-            else:
-                self.file_queue.put((path, url, file_name, size, file_type, full_path))
-                added_file = True
-            processed += 1
-
-        if added_dir and not self.is_processing_dirs:
-            self.window.after(0, self.process_dir_queue)
-        if added_file and not self.is_processing_files:
-            self.window.after(0, self.process_file_queue)
-
-        if reschedule and self.is_scanning and not self.scan_item_buffer.empty():
-            self._schedule_scan_item_flush()
-
-    def on_scan_finished(self, stopped):
-        """Backend hook: finalize UI state after scan ends."""
-        self._run_on_ui_thread(self._on_scan_finished_ui, stopped)
-
-    def _on_scan_finished_ui(self, stopped):
-        self.is_scanning = False
-        self._cancel_scan_item_flush()
-        while not self.scan_item_buffer.empty():
-            self._flush_scan_item_buffer(
-                max_items=self.scan_flush_batch_size,
-                reschedule=False,
-            )
-
-        self.scan_btn.configure(state="normal", text="Scan")
-        self.scan_pause_btn.configure(state="disabled", text="Pause Scan")
-        self.scan_pause_event.set()
-        if not stopped:
-            self.progress_label.configure(text="Scan finished.")
-            self.log_message("[Scan] Finished.")
-        else:
-            self.progress_label.configure(text="Scan stopped.")
-            self.log_message("[Scan] Stopped by user.")
-        self.scan_completed()
-
-    def on_downloads_finished(self, completed, total):
-        """Backend hook: finalize download summary on main thread."""
-        self._run_on_ui_thread(self.downloads_completed, completed, total)
-
-    def process_dir_queue(self):
-        """Processes the directory creation queue to avoid UI blocking."""
-        try:
-            if not hasattr(self, "dir_queue"):
-                return
-            self.is_processing_dirs = True
-            if not self.dir_queue.empty():
-                dir_path, url = self.dir_queue.get()
-                self.create_folder_structure(dir_path, url)
-                self.window.after(10, self.process_dir_queue)
-            else:
-                self.is_processing_dirs = False
-        except tk.TclError as ex:
-            self.is_processing_dirs = False
-            print(f"Error processing directory queue (UI): {str(ex)}")
-
-    def process_file_queue(self):
-        """Processes the file queue to avoid UI blocking."""
-        try:
-            if not hasattr(self, "file_queue"):
-                return
-            self.is_processing_files = True
-            if not self.file_queue.empty():
-                (
-                    dir_path,
-                    url,
-                    file_name,
-                    size,
-                    file_type,
-                    full_path,
-                ) = self.file_queue.get()
-                self.update_file_types(file_name)
-                self._safe_create_folder_and_add_file(
-                    dir_path, url, file_name, size, file_type, full_path
-                )
-                self.window.after(5, self.process_file_queue)
-            else:
-                self.is_processing_files = False
-        except tk.TclError as ex:
-            self.is_processing_files = False
-            print(f"Error processing file queue (UI): {str(ex)}")
-
-    def _safe_create_folder_and_add_file(
-        self, dir_path, url, file_name, size, file_type, full_path
-    ):
-        """Create folder structure and add file in main thread"""
-        try:
-            with self.files_dict_lock:
-                if (
-                    full_path in self.files_dict
-                    and self.files_dict[full_path] is not None
-                ):
-                    return
-
-            parent_id = self.create_folder_structure(dir_path, url)
-
-            for child in self.tree.get_children(parent_id):
-                if "file" in self.tree.item(child)["tags"]:
-                    child_text = self.tree.item(child)["text"]
-                    child_name = child_text.replace(f"{self.file_icon} ", "")
-                    if child_name == file_name:
-                        with self.files_dict_lock:
-                            if full_path in self.files_dict:
-                                del self.files_dict[full_path]
-                        return
-
-            self._add_file_to_tree(
-                parent_id, file_name, size, file_type, full_path, url
-            )
-        except (tk.TclError, OSError) as ex:
-            with self.files_dict_lock:
-                if full_path in self.files_dict:
-                    del self.files_dict[full_path]
-            print(f"Error adding file to tree structure: {str(ex)}")
-
-    def _add_file_to_tree(self, parent_id, file_name, size, file_type, full_path, url):
-        """Add file to tree structure in main thread"""
-        try:
-            items_count = len(self.tree.get_children(parent_id))
-            row_tag = "evenrow" if items_count % 2 == 0 else "oddrow"
-
-            self.tree.insert(
-                parent_id,
-                "end",
-                text=f"{self.file_icon} {file_name}",
-                values=(size, file_type),
-                tags=("file", row_tag),
-            )
-            with self.files_dict_lock:
-                self.files_dict[full_path] = url
-        except tk.TclError as ex:
-            print(f"Error in _add_file_to_tree (TclError): {ex}")
-        except RuntimeError as ex:
-            print(f"Error in _add_file_to_tree (General): {ex}")
-
-    def update_file_types(self, file_name):
-        """Updates the file type filter UI."""
-        ext = normalize_extension(file_name)
-
-        if ext not in self.file_types:
-            self.file_types[ext] = ctk.BooleanVar(value=False)
-            self.file_type_counts[ext] = 0
-            # Re-draw checkboxes
-            self.redraw_file_type_filters()
-        self.file_type_counts[ext] += 1
-        # Update count on the checkbox
-        for child in self.filter_checkboxes_frame.winfo_children():
-            if child.cget("text").startswith(ext):
-                # Use dict expansion to avoid type-checker complaints about unknown kwargs
-                child.configure(**{"text": f"{ext} ({self.file_type_counts[ext]})"})
-                break
-
-    def redraw_file_type_filters(self):
-        """Clears and redraws the file type filter checkboxes."""
-        for widget in self.filter_checkboxes_frame.winfo_children():
-            widget.destroy()
-
-        sorted_types = sorted(self.file_types.keys())
-        for i, file_type in enumerate(sorted_types):
-            var = self.file_types[file_type]
-            count = self.file_type_counts.get(file_type, 0)
-            chk = ctk.CTkCheckBox(
-                self.filter_checkboxes_frame,
-                text=f"{file_type} ({count})",
-                variable=var,
-                command=self.apply_filters,
-                **self.ui_tokens["checkbox"],
-            )
-            chk.grid(row=i // 4, column=i % 4, padx=5, pady=2, sticky="w")
-
-    def apply_filters(self):
-        """Applies selected file type filters by checking/unchecking items in the tree."""
-        all_items = self.get_all_tree_items()
-        items_to_expand = set()
-
-        for item in all_items:
-            tags = self.tree.item(item, "tags")
-            if "file" in tags:
-                text = self.tree.item(item, "text")
-                file_name = text.replace(f"{self.file_icon} ", "")
-
-                ext = normalize_extension(file_name)
-
-                if ext in self.file_types:
-                    should_be_checked = self.file_types[ext].get()
-                    self.toggle_check(item, force_check=should_be_checked)
-                    if should_be_checked:
-                        # If a file is checked, we need to expand its ancestors
-                        parent = self.tree.parent(item)
-                        while parent:
-                            items_to_expand.add(parent)
-                            parent = self.tree.parent(parent)
-
-        for item_id in items_to_expand:
-            self.tree.item(item_id, open=True)
-
-    def get_all_tree_items(self, parent=""):
-        """Recursively gets all item IDs from the treeview."""
-        items = []
-        for child in self.tree.get_children(parent):
-            items.append(child)
-            items.extend(self.get_all_tree_items(child))
-        return items
-
-    def select_all_types(self):
-        """Selects all file type checkboxes."""
-        for var in self.file_types.values():
-            var.set(True)
-        self.apply_filters()
-
-    def deselect_all_types(self):
-        """Deselects all file type checkboxes."""
-        for var in self.file_types.values():
-            var.set(False)
-        self.apply_filters()
-
-    def clear_scan_results(self):
-        """Clears the treeview and all scan-related data."""
-        if self.is_scanning:
-            messagebox.showwarning(
-                "Warning", "Cannot clear results while a scan is in progress."
-            )
-            return
-
-        # Clear Treeview
-        for item in self.tree.get_children():
-            self.tree.delete(item)
-
-        # Clear data structures
-        with self.files_dict_lock:
-            self.files_dict.clear()
-        with self.folders_dict_lock:
-            self.folders.clear()
-
-        self.checked_items.clear()
-        self.file_types.clear()
-        self.file_type_counts.clear()
-
-        # Clear file type filter UI
-        for widget in self.filter_checkboxes_frame.winfo_children():
-            widget.destroy()
-
-        # Reset scan progress
-        self.total_urls = 0
-        self.scanned_urls = 0
-        self.progress_bar.set(0)
-        self.progress_label.configure(text="Scan results cleared.")
-        self.search_var.set("")
-        self.search_entry.configure(state="disabled")
-        self.full_tree_backup.clear()
-
-    def on_search_filter_changed(self, *_):
-        """Callback when the search entry text changes."""
-        self.apply_search_filter()
-
-    def apply_search_filter(self):
-        """Filters the treeview based on the search query."""
-        query = self.search_var.get().lower()
-
-        # Detach all items first
-        for item_id in self.full_tree_backup:
-            if self.tree.exists(item_id):
-                self.tree.detach(item_id)
-
-        if not query:
-            # If query is empty, reattach all items in their original order
-            for item_id, data in self.full_tree_backup.items():
-                if self.tree.exists(item_id):
-                    self.tree.move(item_id, data["parent"], data["index"])
-            return
-
-        items_to_show = set()
-        # Find items that match the query
-        for item_id, data in self.full_tree_backup.items():
-            if query in data["text"].lower():
-                items_to_show.add(item_id)
-                # Add all parents to ensure the path is visible
-                parent = data["parent"]
-                while parent:
-                    items_to_show.add(parent)
-                    parent = self.full_tree_backup[parent]["parent"]
-
-        # Re-attach only the items that should be visible
-        for item_id, data in self.full_tree_backup.items():
-            if item_id in items_to_show:
-                if self.tree.exists(item_id):
-                    self.tree.move(item_id, data["parent"], data["index"])
-                    if "folder" in data["tags"]:
-                        self.tree.item(item_id, open=True)
-
-    def update_thread_count(self, new_count_str: str):
-        """Updates the number of concurrent download threads."""
-        try:
-            new_count = int(new_count_str)
-            if 1 <= new_count <= 10 and new_count != self.max_workers:
-                self.max_workers = new_count
-                self.executor = rebuild_executor(self.executor, self.max_workers)
-        except (ValueError, TypeError):
-            pass
-
-    def update_progress(self, file_path, file_name, progress):
-        """Updates per-file or global progress from the main thread."""
-        self.window.after(
-            0, lambda: self._update_progress_ui(file_path, file_name, progress)
-        )
-
-    def _update_progress_ui(self, file_path, file_name, progress):
-        """Update per-file progress if present; fallback to global bar."""
-        item = self.downloads_panel._items.get(file_path)
-        if item:
-            self.downloads_panel.set_progress(file_path, progress)
-        else:
-            self.progress_bar.set(progress / 100)
-            self.progress_label.configure(
-                text=f"Downloading: {file_name} ({progress:.1f}%)"
-            )
-
-    def update_download_status(self, file_path, status_text):
-        """Update status label for a specific download item."""
-        item = self.downloads_panel._items.get(file_path)
-        if item:
-            self.downloads_panel.set_status(file_path, status_text)
-
-    def log_message(self, message: str):
-        """Append a message to the log panel."""
-        try:
-            self.log_text.configure(state="normal")
-            self.log_text.insert("end", message + "\n")
-            self.log_text.see("end")
-            self.log_text.configure(state="disabled")
-        except tk.TclError:
-            pass
-
-    def download_selected(self):
-        """Download checked files"""
-        if not self.checked_items:
-            messagebox.showwarning("Warning", "Please select files to download")
-            return
-
-        files_to_download = [
-            item
-            for item in self.checked_items
-            if "file" in self.tree.item(item)["tags"]
-        ]
-
-        if not files_to_download:
-            messagebox.showwarning("Warning", "No files selected to download")
-            return
-
-        if not self.download_path:
-            self.choose_download_path()
-            if not self.download_path:  # User cancelled
-                return
-
-        self.pause_event.set()
-        self.pause_btn.configure(state="normal")
-        self.progress_label.configure(text="Preparing to download...")
-
-        futures = []
-        for item in files_to_download:
-            text = self.tree.item(item)["text"]
-            file_name = text.replace(f"{self.file_icon} ", "")
-
-            path_parts = []
-            sanitized_path_parts = []
-            current = item
-            while current:
-                parent = self.tree.parent(current)
-                if not parent:
-                    break
-                parent_text = self.tree.item(parent)["text"]
-                folder_name = parent_text.replace(f"{self.folder_icon} ", "")
-                path_parts.append(folder_name)
-                sanitized_path_parts.append(sanitize_path_segment(folder_name))
-                current = parent
-
-            path_parts.reverse()
-            sanitized_path_parts.reverse()
-            relative_path = os.path.join(*path_parts) if path_parts else ""
-
-            full_path_key = os.path.join(relative_path, file_name).replace("\\", "/")
-            if full_path_key.startswith("/"):
-                full_path_key = full_path_key[1:]
-
-            safe_file_name = sanitize_filename(file_name)
-
-            try:
-                target_dir = safe_join(self.download_path, sanitized_path_parts)
-                file_path = safe_join(self.download_path, sanitized_path_parts + [safe_file_name])
-            except ValueError:
-                self.log_message(f"[Download] Skipped unsafe path: {full_path_key}")
-                continue
-
-            os.makedirs(target_dir, exist_ok=True)
-
-            url = self.files_dict.get(full_path_key)
-
-            if url:
-                cancel_event = self.downloads_panel.ensure(file_path, safe_file_name)
-                # Submit with cancel support and per-file progress
-                future = self.executor.submit(
-                    self.backend.download_file,
-                    url,
-                    file_path,
-                    safe_file_name,
-                    cancel_event,
-                )
-                futures.append(future)
-
-        threading.Thread(target=self.backend.monitor_downloads, args=(futures,)).start()
-
-    def downloads_completed(self, completed, total):
-        """Finalizes the download process and shows a summary."""
-        self.progress_label.configure(
-            text=f"Download completed: {completed}/{total} files."
-        )
-        self.pause_btn.configure(state="disabled")
-        messagebox.showinfo(
-            "Completed",
-            f"{completed} of {total} selected files downloaded successfully.",
-        )
-
-    def start_scan(self):
-        """Starts the website scanning process."""
-        url = self.url_entry.get()
-        if not url:
-            messagebox.showerror("Error", "Please enter URL")
-            return
-
-        if self.is_scanning:
-            self.backend.should_stop = True
-            self.scan_btn.configure(text="Stopping...")
-            self.scan_btn.configure(state="disabled")
-            self._set_status("Stopping", "stopping")
-            return
-
-        self.backend.should_stop = False
-        try:
-            self.download_path = default_download_folder(url, os.getcwd())
-        except (OSError, TypeError) as ex:
-            print(f"Error setting download path: {str(ex)}")
-            self.download_path = os.path.join(os.getcwd(), "downloads")
-
-        self.progress_label.configure(text="Preparing to scan...")
-        self._set_status("Scanning", "scanning")
-        self.scan_btn.configure(text="Stop Scan")
-        self.search_entry.configure(state="disabled")
-        self.search_var.set("")
-        Thread(target=self.backend.scan_website, args=(url,)).start()
-
-    def scan_completed(self):
-        """Called by the backend when scanning is finished."""
-        self.search_entry.configure(state="normal")
-        self.full_tree_backup.clear()
-        for item in self.get_all_tree_items():
-            self.full_tree_backup[item] = {
-                "parent": self.tree.parent(item),
-                "index": self.tree.index(item),
-                "text": self.tree.item(item, "text"),
-                "tags": self.tree.item(item, "tags"),
-            }
-        self._set_status("Ready", "ready")
-
-    def focus_search(self, _event=None):
-        if self.search_entry.cget("state") != "disabled":
-            self.search_entry.focus_set()
-            self.search_entry.icursor("end")
-
-    def focus_logs(self, _event=None):
-        try:
-            if not self.panels_visible:
-                self.panels_tabview.grid()
-                self.panels_visible = True
-                self.toggle_panels_btn.configure(text="Hide Panels")
-            self.panels_tabview.set("Logs")
-            self.log_text.focus_set()
-        except (AttributeError, tk.TclError):
-            pass
-
-    def clear_search(self, _event=None):
-        if self.search_entry.cget("state") != "disabled":
-            self.search_var.set("")
-
-    def run(self):
-        """Starts the Tkinter main loop."""
-        self.window.mainloop()
-
-    def show_context_menu(self, event):
-        """Show right-click menu"""
-        self.context_menu.post(event.x_root, event.y_root)
-
-    def select_all(self):
-        """Select all items"""
-        for item in self.get_all_tree_items():
-            self.toggle_check(item, force_check=True)
-
-    def deselect_all(self):
-        """Deselect all items"""
-        for item in self.get_all_tree_items():
-            self.toggle_check(item, force_check=False)
-
-    def expand_all(self, parent=""):
-        """Expand all nodes in the treeview."""
-        for item in self.tree.get_children(parent):
-            self.tree.item(item, open=True)
-            self.expand_all(item)
-
-    def collapse_all(self, parent=""):
-        """Collapse all nodes in the treeview."""
-        for item in self.tree.get_children(parent):
-            self.tree.item(item, open=False)
-            self.collapse_all(item)
-
-    def on_closing(self):
-        """Handles the window closing event."""
-        # Persist settings on close (best-effort)
-        data = {
-            "panels_visible": bool(getattr(self, "panels_visible", False)),
-            "active_tab": str(self.panels_tabview.get()),
-        }
-        settings_path = getattr(
-            self,
-            "_settings_path",
-            os.path.join(os.path.expanduser("~"), ".index_ripper_settings.json"),
-        )
-        save_settings(settings_path, data)
-        if messagebox.askokcancel("Quit", "Do you want to quit?"):
-            self.backend.should_stop = True
-            self.executor.shutdown(wait=False, cancel_futures=True)
-            self.window.destroy()
-
-    def sort_tree(self, col):
-        """Sorts the treeview columns."""
-        items = [(self.tree.set(i, col), i) for i in self.tree.get_children("")]
-
-        # A simple string sort
-        items.sort(reverse=self.sort_reverse)
-
-        for index, (_, i) in enumerate(items):
-            self.tree.move(i, "", index)
-
-        self.sort_reverse = not self.sort_reverse
-
-
-if __name__ == "__main__":
-    if "--ui-smoke" in sys.argv:
-        from ui_ctk import WebsiteCopierCtk
-        app = WebsiteCopierCtk(ui_smoke=True)
-        app.window.after(0, app.window.destroy)
-        app.run()
-        print("UI_SMOKE_OK")
-        raise SystemExit(0)
-
-    from ui_ctk import WebsiteCopierCtk
-    app = WebsiteCopierCtk()
-    app.run()
