+++ src/index_ripper/__init__.py
@@ -1,0 +1,3 @@
+"""Index Ripper — download files from directory listing pages."""
+
+__version__ = "0.1.0"
