+++ src/index_ripper/self_test.py
