+++ src/index_ripper/settings.py
