+++ src/index_ripper/ui/__init__.py
@@ -1,0 +1,1 @@
+"""Index Ripper UI components."""
