+++ src/index_ripper/utils.py
