+++ tests/test_theme.py
@@ -1,22 +1,22 @@
 import unittest
 
-import ui_theme
+from index_ripper.ui import theme
 
 
 class TestUITheme(unittest.TestCase):
     def test_action_button_style_name(self):
-        self.assertEqual(ui_theme.action_button_style_name("primary"), "Primary.TButton")
+        self.assertEqual(theme.action_button_style_name("primary"), "Primary.TButton")
         self.assertEqual(
-            ui_theme.action_button_style_name("secondary"), "Secondary.TButton"
+            theme.action_button_style_name("secondary"), "Secondary.TButton"
         )
-        self.assertEqual(ui_theme.action_button_style_name("danger"), "Danger.TButton")
+        self.assertEqual(theme.action_button_style_name("danger"), "Danger.TButton")
         self.assertEqual(
-            ui_theme.action_button_style_name("success"), "Success.TButton"
+            theme.action_button_style_name("success"), "Success.TButton"
         )
-        self.assertEqual(ui_theme.action_button_style_name("unknown"), "Secondary.TButton")
+        self.assertEqual(theme.action_button_style_name("unknown"), "Secondary.TButton")
 
     def test_ui_tokens_contains_modern_widget_styles(self):
-        tokens = ui_theme.ui_tokens()
+        tokens = theme.ui_tokens()
         self.assertIn("option_menu", tokens)
         self.assertIn("tabview", tokens)
         self.assertIn("checkbox", tokens)
@@ -25,7 +25,7 @@ class TestUITheme(unittest.TestCase):
         self.assertIn("border_color", tokens["checkbox"])
 
     def test_apply_bootstrap_theme_not_present(self):
-        self.assertFalse(hasattr(ui_theme, "apply_bootstrap_theme"))
+        self.assertFalse(hasattr(theme, "apply_bootstrap_theme"))
 
 
 if __name__ == "__main__":
