+++ tests/test_utils.py
@@ -2,12 +2,31 @@ import os
 import tempfile
 import unittest
 
-from app_utils import (
+from index_ripper.utils import (
+    default_download_folder,
     is_url_in_scope,
+    normalize_extension,
     safe_join,
     sanitize_filename,
     sanitize_path_segment,
+    shorten_path,
 )
+
+
+class TestAppUtils(unittest.TestCase):
+    def test_normalize_extension_normal(self):
+        self.assertEqual(normalize_extension("a.TXT"), ".txt")
+
+    def test_normalize_extension_none(self):
+        self.assertEqual(normalize_extension("README"), ".(No Extension)")
+
+    def test_shorten_path(self):
+        self.assertEqual(shorten_path("/a/b", keep=10), "/a/b")
+        self.assertEqual(shorten_path("/" + "x" * 50, keep=10), "..." + "x" * 10)
+
+    def test_default_download_folder(self):
+        out = default_download_folder("https://example.com/a/b", "/tmp")
+        self.assertTrue(out.endswith("example.com"))
 
 
 class TestSecurityUtils(unittest.TestCase):
